requirements.txt
requests==2.28.1
boto3==1.26.0
schedule==1.1.0
python-dotenv==0.21.1
